# backend/tests/snapshots/openapi_schema.json
{
  "components": {
    "schemas": {
...
      "AdminDeleteUserResponse": {
        "description": "Response for DELETE /admin/users/{user_id}.",
        "properties": {
          "deleted": {
            "title": "Deleted",
            "type": "boolean"
          },
          "user_id": {
            "title": "User Id",
            "type": "string"
          }
        },
        "required": [
          "deleted",
          "user_id"
        ],
        "title": "AdminDeleteUserResponse",
        "type": "object"
      },
...
      "AdminUpdateUserResponse": {
        "description": "Response for PUT /admin/users/{user_id}.",
        "properties": {
          "updated": {
            "title": "Updated",
            "type": "boolean"
          },
          "user_id": {
            "title": "User Id",
            "type": "string"
          }
        },
        "required": [
          "updated",
          "user_id"
        ],
        "title": "AdminUpdateUserResponse",
        "type": "object"
      },
...
      "DeleteAccountResponse": {
        "description": "Response for DELETE /me.",
        "properties": {
          "message": {
            "title": "Message",
            "type": "string"
          },
          "success": {
            "title": "Success",
            "type": "boolean"
          }
        },
        "required": [
          "success",
          "message"
        ],
        "title": "DeleteAccountResponse",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "SuccessResponse": {
        "description": "Generic success response for simple operations.",
        "properties": {
          "success": {
            "title": "Success",
            "type": "boolean"
          }
        },
        "required": [
          "success"
        ],
        "title": "SuccessResponse",
        "type": "object"
      },
...
      "UpdateUserRequest": {
        "properties": {
          "company": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Company"
          },
          "full_name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Full Name"
          },
          "plan_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Plan Id"
          }
        },
        "title": "UpdateUserRequest",
        "type": "object"
      },
...
      "UserProfileResponse": {
        "description": "User profile with plan capabilities and quota status.\n\nReturned by /api/me endpoint to provide frontend with all\nnecessary plan information for UI rendering.",
        "properties": {
          "capabilities": {
            "additionalProperties": true,
            "description": "Plan capabilities (max_history_days, allow_excel, etc.)",
            "title": "Capabilities",
            "type": "object"
          },
          "email": {
            "title": "Email",
            "type": "string"
          },
          "is_admin": {
            "default": false,
            "description": "Whether user has admin privileges",
            "title": "Is Admin",
            "type": "boolean"
          },
          "plan_id": {
            "description": "Plan ID (e.g., 'consultor_agil')",
            "title": "Plan Id",
            "type": "string"
          },
          "plan_name": {
            "description": "Display name (e.g., 'Consultor \u00c1gil')",
            "title": "Plan Name",
            "type": "string"
          },
          "quota_remaining": {
            "description": "Searches remaining this month",
            "minimum": 0.0,
            "title": "Quota Remaining",
            "type": "integer"
          },
          "quota_reset_date": {
            "description": "ISO 8601 timestamp of next quota reset",
            "title": "Quota Reset Date",
            "type": "string"
          },
          "quota_used": {
            "description": "Searches used this month",
            "minimum": 0.0,
            "title": "Quota Used",
            "type": "integer"
          },
          "subscription_status": {
            "description": "Status: 'trial', 'active', or 'expired'",
            "title": "Subscription Status",
            "type": "string"
          },
          "trial_expires_at": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "description": "ISO 8601 timestamp when trial expires (if applicable)",
            "title": "Trial Expires At"
          },
          "user_id": {
            "title": "User Id",
            "type": "string"
          }
        },
        "required": [
          "user_id",
          "email",
          "plan_id",
          "plan_name",
          "capabilities",
          "quota_used",
          "quota_remaining",
          "quota_reset_date",
          "subscription_status"
        ],
        "title": "UserProfileResponse",
        "type": "object"
      },
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      },
...
    },
    "securitySchemes": {
      "HTTPBearer": {
        "scheme": "bearer",
        "type": "http"
      }
    }
  },
  "info": {
    "description": "API para busca e an\u00e1lise de licita\u00e7\u00f5es de uniformes no PNCP. Permite filtrar oportunidades por estado, valor e keywords, gerando relat\u00f3rios Excel e resumos executivos via IA.",
    "title": "BidIQ Uniformes API",
    "version": "0.2.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/admin/users/{user_id}": {
      "delete": {
        "description": "Delete a user and all their data.",
        "operationId": "delete_user_admin_users__user_id__delete",
        "parameters": [
          {
            "description": "User UUID to delete",
            "in": "path",
            "name": "user_id",
            "required": true,
            "schema": {
              "description": "User UUID to delete",
              "title": "User Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AdminDeleteUserResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Delete User",
        "tags": [
          "admin"
        ]
      },
      "put": {
        "description": "Update user profile or plan.",
        "operationId": "update_user_admin_users__user_id__put",
        "parameters": [
          {
            "description": "User UUID to update",
            "in": "path",
            "name": "user_id",
            "required": true,
            "schema": {
              "description": "User UUID to update",
              "title": "User Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateUserRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AdminUpdateUserResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Update User",
        "tags": [
          "admin"
        ]
      }
    },
...
    "/change-password": {
      "post": {
        "description": "Change current user's password.",
        "operationId": "change_password_change_password_post",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SuccessResponse"
                }
              }
            },
            "description": "Successful Response"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Change Password",
        "tags": [
          "user"
        ]
      }
    },
...
    "/me": {
      "delete": {
        "description": "Delete entire user account and all associated data (LGPD Art. 18 VI).\n\nCascade order:\n1. Cancel active Stripe subscription (if any)\n2. Delete from: search_sessions, monthly_quota, user_subscriptions,\n   user_oauth_tokens, messages, profiles\n3. Delete auth user via Supabase admin API\n4. Log anonymized audit entry",
        "operationId": "delete_account_me_delete",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DeleteAccountResponse"
                }
              }
            },
            "description": "Successful Response"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Delete Account",
        "tags": [
          "user"
        ]
      },
      "get": {
        "description": "Get current user profile with plan capabilities and quota status.",
        "operationId": "get_profile_me_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserProfileResponse"
                }
              }
            },
            "description": "Successful Response"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Get Profile",
        "tags": [
          "user"
        ]
      }
    },
    "/me/export": {
      "get": {
        "description": "Export all user data as JSON file (LGPD Art. 18 V \u2014 data portability).\n\nReturns a downloadable JSON file containing:\n- Profile information\n- Search history (sessions)\n- Subscription history\n- Messages",
        "operationId": "export_user_data_me_export_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Export User Data",
        "tags": [
          "user"
        ]
      }
    },
...
    "/v1/admin/users/{user_id}": {
      "delete": {
        "description": "Delete a user and all their data.",
        "operationId": "delete_user_v1_admin_users__user_id__delete",
        "parameters": [
          {
            "description": "User UUID to delete",
            "in": "path",
            "name": "user_id",
            "required": true,
            "schema": {
              "description": "User UUID to delete",
              "title": "User Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AdminDeleteUserResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Delete User",
        "tags": [
          "admin"
        ]
      },
      "put": {
        "description": "Update user profile or plan.",
        "operationId": "update_user_v1_admin_users__user_id__put",
        "parameters": [
          {
            "description": "User UUID to update",
            "in": "path",
            "name": "user_id",
            "required": true,
            "schema": {
              "description": "User UUID to update",
              "title": "User Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateUserRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AdminUpdateUserResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Update User",
        "tags": [
          "admin"
        ]
      }
    },
...
    "/v1/change-password": {
      "post": {
        "description": "Change current user's password.",
        "operationId": "change_password_v1_change_password_post",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SuccessResponse"
                }
              }
            },
            "description": "Successful Response"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Change Password",
        "tags": [
          "user"
        ]
      }
    },
...
    "/v1/me": {
      "delete": {
        "description": "Delete entire user account and all associated data (LGPD Art. 18 VI).\n\nCascade order:\n1. Cancel active Stripe subscription (if any)\n2. Delete from: search_sessions, monthly_quota, user_subscriptions,\n   user_oauth_tokens, messages, profiles\n3. Delete auth user via Supabase admin API\n4. Log anonymized audit entry",
        "operationId": "delete_account_v1_me_delete",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DeleteAccountResponse"
                }
              }
            },
            "description": "Successful Response"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Delete Account",
        "tags": [
          "user"
        ]
      },
      "get": {
        "description": "Get current user profile with plan capabilities and quota status.",
        "operationId": "get_profile_v1_me_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserProfileResponse"
                }
              }
            },
            "description": "Successful Response"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Get Profile",
        "tags": [
          "user"
        ]
      }
    },
    "/v1/me/export": {
      "get": {
        "description": "Export all user data as JSON file (LGPD Art. 18 V \u2014 data portability).\n\nReturns a downloadable JSON file containing:\n- Profile information\n- Search history (sessions)\n- Subscription history\n- Messages",
        "operationId": "export_user_data_v1_me_export_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Export User Data",
        "tags": [
          "user"
        ]
      }
    },
...
  }
}